fix: resolve mock portal hydration error on staging
